Support subclasses by passing type-constraints with model
There were two alternatives to passing type-constraints to `walk` so
that all functions would work with subclass constraints:
- Pass as optional argument, which would need to be added to every
function dependent (directly or indirectly) on `walk`
- Pass assoc'ed to model
The latter approach was chosen as it would require much less refactoring
in Bluegenes, the downside being it's more subtle. Hopefully mentioning
this in every docstring will be enough.