Make deconstruct-by-class match im-tables behaviour
Due to how the view is overwritten by a single target class as required
by tolist WS, this leads to a differing query result set, due to views
implicitly constraining refs/colls to be present (as inner join is the
default). To match this behaviour without views, we need to add IS NOT
NULL constraints for them, except in the case where the class is an
outer join, when it should not receive a substitute constraint.